datafeeder: refactoring, use JobManager to launch spring-batch jobs asynchronously
