+++ datafeeder/src/main/java/org/georchestra/datafeeder/batch/analysis/DataUploadAnalysisService.java
@@ -18,9 +18,6 @@
  */
 package org.georchestra.datafeeder.service.batch.analysis;
 
-import static org.georchestra.datafeeder.service.batch.analysis.UploadAnalysisConfiguration.JOB_NAME;
-import static org.georchestra.datafeeder.service.batch.analysis.UploadAnalysisConfiguration.UPLOAD_ID_JOB_PARAM_NAME;
-
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.nio.file.Path;
@@ -44,19 +41,10 @@ import org.georchestra.datafeeder.service.DatasetMetadata;
 import org.georchestra.datafeeder.service.DatasetsService;
 import org.georchestra.datafeeder.service.FileStorageService;
 import org.georchestra.datafeeder.service.UploadPackage;
+import org.georchestra.datafeeder.service.batch.JobManager;
 import org.geotools.util.Converters;
 import org.locationtech.jts.geom.Geometry;
-import org.springframework.batch.core.Job;
-import org.springframework.batch.core.JobExecution;
-import org.springframework.batch.core.JobParameters;
-import org.springframework.batch.core.JobParametersBuilder;
-import org.springframework.batch.core.JobParametersInvalidException;
-import org.springframework.batch.core.launch.JobLauncher;
-import org.springframework.batch.core.repository.JobExecutionAlreadyRunningException;
-import org.springframework.batch.core.repository.JobInstanceAlreadyCompleteException;
-import org.springframework.batch.core.repository.JobRestartException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
 import lombok.NonNull;
@@ -71,13 +59,12 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 public class DataUploadAnalysisService {
 
+    private @Autowired JobManager jobManager;
+
     private @Autowired @Setter FileStorageService fileStore;
     private @Autowired DataUploadJobRepository jobRepository;
     private @Autowired DatasetUploadStateRepository datasetRepository;
     private @Autowired DatasetsService datasetsService;
-
-    private @Autowired JobLauncher jobLauncher;
-    private @Autowired @Qualifier(JOB_NAME) Job uploadAnalysisJob;
 
     /**
      * Data upload analysis process step 0: creates a {@link DataUploadJob} with
@@ -103,23 +90,7 @@ public class DataUploadAnalysisService {
     }
 
     public void runJob(@NonNull UUID jobId) {
-        final String paramName = UPLOAD_ID_JOB_PARAM_NAME;
-        final String paramValue = jobId.toString();
-        final boolean identifying = true;
-
-        final JobParameters params = new JobParametersBuilder()//
-                .addString(paramName, paramValue, identifying)//
-                .toJobParameters();
-        log.info("Launching analisys job {}", jobId);
-        JobExecution execution;
-        try {
-            execution = jobLauncher.run(uploadAnalysisJob, params);
-        } catch (JobExecutionAlreadyRunningException | JobRestartException | JobInstanceAlreadyCompleteException
-                | JobParametersInvalidException e) {
-            log.error("Error running job {}", jobId, e);
-            throw new RuntimeException("Error running job " + jobId, e);
-        }
-        log.info("Analysis job {} finished with status {}", jobId, execution.getStatus());
+        jobManager.launchUploadJobAnalysis(jobId);
     }
 
     /**
