+++ datafeeder/src/main/java/org/georchestra/datafeeder/batch/analysis/UploadAnalysisJobConfiguration.java
