+++ datafeeder/src/main/java/org/georchestra/datafeeder/batch/publish/DataPublishingJobConfiguration.java
