+++ datafeeder/src/main/java/org/georchestra/datafeeder/service/DataFeederServiceConfiguration.java
@@ -33,12 +33,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
-import org.springframework.scheduling.annotation.EnableAsync;
 
 import lombok.extern.slf4j.Slf4j;
 
 @Configuration
-@EnableAsync
 @Import(value = { PersistenceConfiguration.class, DatafeederBatchConfiguration.class })
 @Slf4j
 public class DataFeederServiceConfiguration {
