--- datafeeder/src/main/java/org/georchestra/datafeeder/service/batch/DatafeederBatchConfiguration.java
@@ -1,37 +1,0 @@
-/*
- * Copyright (C) 2020 by the geOrchestra PSC
- *
- * This file is part of geOrchestra.
- *
- * geOrchestra is free software: you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free
- * Software Foundation, either version 3 of the License, or (at your option)
- * any later version.
- *
- * geOrchestra is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
- * more details.
- *
- * You should have received a copy of the GNU General Public License along with
- * geOrchestra.  If not, see <http://www.gnu.org/licenses/>.
- */
-package org.georchestra.datafeeder.service.batch;
-
-import org.georchestra.datafeeder.service.batch.analysis.UploadAnalysisConfiguration;
-import org.georchestra.datafeeder.service.batch.publish.DataPublishingConfiguration;
-import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-
-@Configuration
-@EnableBatchProcessing
-@Import({ UploadAnalysisConfiguration.class, DataPublishingConfiguration.class })
-public class DatafeederBatchConfiguration {
-
-    public @Bean JobManager jobManager() {
-        return new JobManager();
-    }
-
-}
