+++ datafeeder/src/test/java/org/georchestra/datafeeder/batch/analysis/BatchTestConfiguration.java
