+++ datafeeder/src/test/java/org/georchestra/datafeeder/batch/analysis/UploadAnalysisJobConfigurationTest.java
